start qa module
